Super hacky proof of concept for kubernetes external mounts
This is a quickly written, poorly organized hack that demonstrates
the basic outline of how I currently plan to implement external volume
support on kubernetes. This currently supports host directory mounts and
git repositories, as they did not require any api interaction.

This is not expected to be merged without serious modifications,
it merely serves as proof of concept.